pyproject.toml
# Python project configuration for Paper Mimic
# This file configures Python project metadata

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-mimic"
version = "1.0.0"
description = "Intelligent Question Paper Mimicker - Generate similar questions from reference exam papers"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

# Dependencies
dependencies = [
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "tiktoken>=0.5.0",
    "requests>=2.32.2",
    "openai>=1.30.0",
    "aiohttp>=3.9.4",
    "httpx>=0.27.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "magic-pdf[full]>=0.1.0",
]
